debug hinge loss

+++ main.py
@@ -28,7 +28,7 @@ def loss_func(args, f, y):
 
 def loss_func_prime(args, f, y):
     if args.loss == 'hinge':
-        return -((args.loss_margin - args.alpha * f * y) > 0) * y
+        return -((args.loss_margin - args.alpha * f * y) > 0).double() * y
     if args.loss == 'softhinge':
         return -torch.sigmoid(args.loss_beta * (args.loss_margin - args.alpha * f * y)) * y
     if args.loss == 'qhinge':
